Split spinlock types into multiple files for easier integration to console repos (#6221)
